+++ src/main/java/jarvis/command/TagCommand.java
@@ -1,6 +1,6 @@
 package jarvis.command;
 
-import jarvis.ErrorMessage;
+import jarvis.ui.ErrorMessage;
 import jarvis.tag.Tag;
 import jarvis.task.TaskList;
 import jarvis.task.Task;
