+++ src/main/java/jarvis/command/UntagCommand.java
@@ -1,7 +1,6 @@
 package jarvis.command;
 
-import jarvis.ErrorMessage;
-import jarvis.tag.Tag;
+import jarvis.ui.ErrorMessage;
 import jarvis.task.TaskList;
 import jarvis.task.Task;
