+++ src/main/java/jarvis/ui/Parser.java
@@ -1,4 +1,4 @@
-package jarvis;
+package jarvis.ui;
 
 import jarvis.task.Task;
